pyproject.toml
[tool.poetry]
name = "neo4j-uploader"
version = "0.5.0"
description = ""
authors = ["Jason Koo <jalakoo@gmail.com>"]
readme = "README.md"
packages = [{include = "neo4j_uploader"}]

[tool.poetry.dependencies]
python = "^3.9"
neo4j = "^5.14.1"
jsonschema = "^4.20.0"
cymple = "^0.11.0"
pydantic = "^2.5.2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
